# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.SwaggerError:
    properties:
      code:
        example: VALIDATION_FAILED
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
      requestId:
        example: req-abc123
        type: string
      timestamp:
        example: "2024-02-01T10:30:00Z"
        type: string
    type: object
  internal_handler.SwaggerErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/internal_handler.SwaggerError'
    type: object
  internal_handler.SwaggerExtractImageRequest:
    description: Extract recipe from image request
    properties:
      imageBase64:
        example: /9j/4AAQSkZJRgABAQ...
        type: string
      mimeType:
        example: image/jpeg
        type: string
      saveAuto:
        example: true
        type: boolean
    required:
    - imageBase64
    type: object
  internal_handler.SwaggerExtractImageResponse:
    description: Image extraction response
    properties:
      recipe:
        $ref: '#/definitions/internal_handler.SwaggerExtractionResult'
      savedId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      source:
        example: image
        type: string
    type: object
  internal_handler.SwaggerExtractURLRequest:
    description: Extract recipe from URL request
    properties:
      saveAuto:
        example: true
        type: boolean
      url:
        example: https://example.com/recipe/carbonara
        type: string
    required:
    - url
    type: object
  internal_handler.SwaggerExtractURLResponse:
    description: URL extraction response
    properties:
      alreadyExists:
        example: false
        type: boolean
      recipe:
        $ref: '#/definitions/internal_handler.SwaggerExtractionResult'
      savedId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      source:
        example: webpage
        type: string
    type: object
  internal_handler.SwaggerExtractedIngredient:
    description: AI-extracted ingredient
    properties:
      category:
        example: pantry
        type: string
      isOptional:
        example: false
        type: boolean
      name:
        example: Spaghetti
        type: string
      notes:
        example: dried pasta
        type: string
      quantity:
        example: "400"
        type: string
      unit:
        example: g
        type: string
      videoTimestamp:
        example: 1.5
        type: number
    type: object
  internal_handler.SwaggerExtractedStep:
    description: AI-extracted recipe step
    properties:
      durationSeconds:
        example: 600
        type: integer
      instruction:
        example: Boil pasta according to package directions
        type: string
      stepNumber:
        example: 1
        type: integer
      technique:
        example: boiling
        type: string
      temperature:
        example: ""
        type: string
      videoTimestampEnd:
        example: 2
        type: number
      videoTimestampStart:
        example: 0.5
        type: number
    type: object
  internal_handler.SwaggerExtractionResult:
    description: AI-extracted recipe data
    properties:
      cookTime:
        example: 20
        type: integer
      cuisine:
        example: Italian
        type: string
      description:
        example: Classic Roman pasta dish
        type: string
      difficulty:
        example: medium
        type: string
      ingredients:
        items:
          $ref: '#/definitions/internal_handler.SwaggerExtractedIngredient'
        type: array
      prepTime:
        example: 10
        type: integer
      servings:
        example: 4
        type: integer
      steps:
        items:
          $ref: '#/definitions/internal_handler.SwaggerExtractedStep'
        type: array
      tags:
        example:
        - pasta
        - italian
        - quick
        items:
          type: string
        type: array
      thumbnailUrl:
        example: https://example.com/thumb.jpg
        type: string
      title:
        example: Spaghetti Carbonara
        type: string
    type: object
...
host: api.dishflow.app
info:
  contact:
    email: support@dishflow.app
    name: DishFlow Support
  description: Recipe management API with AI-powered extraction, pantry tracking,
    and shopping lists
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://dishflow.app/terms
  title: DishFlow API
  version: "1.0"
paths:
...
  /recipes/extract-image:
    post:
      consumes:
      - multipart/form-data
      - application/json
      description: Extract a recipe from an uploaded image using AI (cookbook photo,
        screenshot)
      parameters:
      - description: Image file (multipart)
        in: formData
        name: image
        type: file
      - description: Auto-save extracted recipe
        in: formData
        name: saveAuto
        type: boolean
      - description: JSON with base64 image
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handler.SwaggerExtractImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Extracted recipe
          schema:
            $ref: '#/definitions/internal_handler.SwaggerExtractImageResponse'
        "400":
          description: Invalid image
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "422":
          description: Extraction failed or no recipe found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Extract recipe from image
      tags:
      - Recipes
  /recipes/extract-url:
    post:
      consumes:
      - application/json
      description: Extract a recipe from a webpage URL using AI (recipe blogs, cooking
        sites)
      parameters:
      - description: URL and options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerExtractURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Extracted recipe
          schema:
            $ref: '#/definitions/internal_handler.SwaggerExtractURLResponse'
        "400":
          description: Invalid URL
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "422":
          description: Extraction failed or no recipe found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Extract recipe from URL
      tags:
      - Recipes
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...